# Parse huge integer faster