… src/util/pm_string.c
… templates/ext/prism/api_node.c.erb
… templates/src/serialize.c.erb