Parsing huge integer literal is a bit slow.
`Prism.parse('1'*1000000)` takes about 67.8s, `eval('1'*1000000)` takes only 0.32s. This pull request improves it to 1.46s.

Constructing a bigint with `digits.each{v=v*10+_1}` is `O(n^2)`.
To make it faster, I think we need a fast bigint multiplication algorithm.
(I'm not sure this is worth introducing complexity)

### Bigint and multiplication
Implement karatsuba algorithm that ruby's Bignum uses internally.
Implement radix conversion using fast multiplication.

### Parse digits
If base is power of two, just pack bits to bigint.
If base is 10, peform radix conversion.

### pm_integer_t to ::Integer
`"0x...".to_i`  is  faster than `|` `<<` for large integer.

### pm_intger_t to string
Use radix conversion

### Comparison

| Prism.parse(code) | before | after | ratio |
| --- | --- | --- | --- |
| '1'*1000000 | 67.798872s | 1.462340s | x46.3 |
| '1'*100000 | 0.723387s | 0.061709s | x11.7 |
| '0x'+'1'*100000 | 0.908699s | 0.002813s | x323 |
| (['1'*100]*100000).join(';') | 0.322435s | 0.189371s | x1.70 |
| (['1'*20]*100000).join(';') | 0.084510s | 0.088816s | x0.95 slower |